Big news: Minesweeper now has a main menu window with new options and a standard game difficulty !
Images:
A new set of minesweeper face when the number of squares is odd
Added game icon
Added game logo image
Added menubuttons sprite (just for a while)
Added optionbuttons sprite (menubuttons will be here soon)
In main tiles I just change a color

Sounds:
I added a main menu music

Header:
Change the MAX to 100 (more relevant)
Added faceSize (minesweeper face will change if the width is odd or
even)
Added gameTime (from now the game have a timer)
Added menustage (to keep track of the menu stage)
From now width and height have a preset value.

Added music_sound, sfx_sound and mark_flag options in case if you do not
want them.
Added display_size to keep track when I need to resize the game window.

Added a new function "show_surroundings". It's a classic minesweeper
feature, I don't know how to explain it, you'll see.

And I moved all the functions here for more free space in main file

Main:

I implemented the switch statement because I was forced to do so in
order to make the menu stage in game and because of this I had to change
a large part of code
I added the main menu sound
I added the game icon :D
I added the logo, menubuttons and the option buttons sprite
Added the main menu options, Play, Options and Exit.

In the main loop I have 4 cases: Case 0 is the main menu, Case 1 is
actually the game, Case 2 is the options menu and the last Case it is
connected to the case 1. In case 3 you select the game difficulty.

there is more work..